js/inject-anchors: Refactor and enable blink on 'hashchange'
This also changes the behavior of the animation to span the entire block, rather
than rewriting the block text to a span, and then animating the span. This is an intentional
stylistic change since I felt the previous method was not obvious enough. This also fixes
animations 'eating' the anchor link